BattleCity.Stage.S26 (battle_city v0.1.8)

Stage: 26 Difficulty: 2 Bots: ["6fast", "6armor", "4basic", "4power"]

X X R R X X X X X X X X X
Tc X X R F X B5 X X X X X X
F Tc X X X X T5 X B5 X R R X
F F X T3 Bc Ba X X T5 F R X X
F F F X X Tf Bc Ba X X X X Tc F F T3 Tc X Ba X Tf Bc X X Tc F
F T3 X X B3 Tf X Ba X T3 X F F
X X X X X B5 B3 Tf X X F F F
X X R F Ta X X B5 B3 Tc T3 F F
X R R X Ba X Ta X X X X X F
X X X X X X Ba X F R X X T3 Tf X X X X B8 Bc B4 X R R X X
Tf Tf X X X Ba E B5 X X X X Tf

Functions

init(map \\ %{})

Specs

init(map()) :: BattleCity.Stage.t()